created DataReaderService
