+++ Octopus.Core.Loader.BusinessLogic/Services/MessageHandler.cs
@@ -1,40 +1,26 @@
 ﻿using Octopus.Core.Common.DynamicObject.Services.Interfaces;
 using Octopus.Core.Common.Helpers.JsonDeserializer;
 using Octopus.Core.Common.Models;
+using Octopus.Core.Loader.BusinessLogic.Interfaces;
 using Octopus.Core.RabbitMq.Services.Interfaces;
-using System.Collections.Generic;
 using System.Text.Json;
-using System.Threading.Tasks;
 
 namespace Octopus.Core.Loader.BusinessLogic.Services
 {
     public class MessageHandler : IEventProcessor
     {
         private IEntityDescription _entityDescription;
-        private readonly IDynamicObjectCreateService _dynamicObjectCreate;
-        private readonly IJsonDeserializer _jsonDeserializer;
+        private readonly IDataReaderService _dataReaderService;
 
-        public MessageHandler(IDynamicObjectCreateService dynamicObjectCreate, 
-            IJsonDeserializer jsonDeserializer) 
+        public MessageHandler(IDataReaderService dataReaderService)
         {
-            _dynamicObjectCreate = dynamicObjectCreate;
-            _jsonDeserializer = jsonDeserializer;
+            _dataReaderService = dataReaderService;
         }
 
         public async void ProcessEvent(string message)
         {
             _entityDescription = JsonSerializer.Deserialize<EntityDescription>(message);
-
-            var objects = await GetDataFromFileAsync(_entityDescription.EntityFilePath);
-        }
-
-        private async Task<IEnumerable<object>> GetDataFromFileAsync(string filePath)
-        {
-            var typeListOf = typeof(List<>);
-            var extendedType = _dynamicObjectCreate.CreateTypeByDescription();
-            var typeListOfExtendedType = typeListOf.MakeGenericType(extendedType);
-
-            return await _jsonDeserializer.GetDynamicObjects(typeListOfExtendedType, _entityDescription.EntityFilePath);
+            var objects = await _dataReaderService.GetDataFromFileAsync(_entityDescription.EntityFilePath);
         }
     }
 }
